Developer

+++ HomeCook/Models/Ingredient.swift
@@ -1,8 +1,5 @@
 
-import Foundation
-
-
 struct Ingredient {
     let title: String
-    let price: Double
+    let price: Int
 }
